requirements.txt
flask
pymupdf
pytesseract
sentence-transformers
faiss-cpu
google-generativeai
transformers
